REL-2412 Consolidate edu.gemini.gsa packages into a single one time 30m
