+++ bundle/edu.gemini.dataman.app/src/test/scala/edu/gemini/dataman/app/TestSupport.scala
@@ -1,6 +1,6 @@
 package edu.gemini.dataman.app
 
-import edu.gemini.gsa.core.Arbitraries
+import edu.gemini.gsa.query.Arbitraries
 import edu.gemini.pot.sp.{ISPFactory, ISPProgram, ProgramGen}
 import edu.gemini.pot.spdb.{DBLocalDatabase, IDBDatabaseService}
 import edu.gemini.spModel.dataset.{DatasetRecord, DatasetGsaState, SummitState, DatasetQaState, Dataset, DatasetExecRecord, DatasetQaRecord, DatasetLabel}
