+++ bundle/edu.gemini.gsa.query/src/main/scala/edu/gemini/gsa/query/package.scala
@@ -3,8 +3,6 @@ package edu.gemini.gsa
 import java.util.concurrent.atomic.AtomicReference
 import java.util.logging.Level
 
-//import edu.gemini.dataman.core.{DmanAction, DmanFailure}
-
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
